refactor: split pkg/runtime/docker/docker.go by responsibility (issue 3)
Final phase D of the docker.go split: extract stress.go, cgroup.go, pull.go;
delete the now-empty docker.go shell. Move stressContainerCommand tests into
stress_test.go, factor shared test helpers into helpers_test.go, and consolidate
HTTP client tests in http_client_test.go.

Production files in pkg/runtime/docker/ now top out at 299 LOC (netem.go),
well under the 350 LOC modularity target. No public API changes; podman
embed-and-override path unaffected.
