+++ pkg/runtime/docker/stress.go
@@ -3,31 +3,14 @@ package docker
 import (
 	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
-	"strings"
 	"time"
 
 	ctr "github.com/alexei-led/pumba/pkg/container"
 	ctypes "github.com/docker/docker/api/types/container"
-	imagetypes "github.com/docker/docker/api/types/image"
 	log "github.com/sirupsen/logrus"
 )
-
-type imagePullResponse struct {
-	Status         string `json:"status"`
-	Error          string `json:"error"`
-	Progress       string `json:"progress"`
-	ProgressDetail struct {
-		Current int `json:"current"`
-		Total   int `json:"total"`
-	} `json:"progressDetail"`
-}
-
-const cgroupDriverSystemd = "systemd"
-const cgroupDriverCgroupfs = "cgroupfs"
 
 // StressContainer starts stress test on a container (CPU, memory, network, io)
 func (client dockerClient) StressContainer(ctx context.Context, c *ctr.Container, stressors []string, img string, pull bool, duration time.Duration, injectCgroup, dryrun bool) (string, <-chan string, <-chan error, error) {
@@ -45,54 +28,6 @@ func (client dockerClient) StressContainer(ctx context.Context, c *ctr.Container
 		return "", nil, nil, nil
 	}
 	return client.stressContainerCommand(ctx, c.ID(), stressors, img, pull, injectCgroup)
-}
-
-// cgroupDriver queries the Docker daemon for its cgroup driver.
-// Returns the driver name or empty string with error on failure.
-func (client dockerClient) cgroupDriver(ctx context.Context) (string, error) {
-	info, err := client.systemAPI.Info(ctx)
-	if err != nil {
-		log.WithError(err).Warn("failed to get docker info, assuming cgroupfs driver")
-		return "", err
-	}
-	return info.CgroupDriver, nil
-}
-
-// containerLeafCgroup returns the leaf cgroup directory name for a container
-// based on the cgroup driver. On cgroupfs the leaf is the container ID; on
-// systemd it is a scope unit named "docker-<id>.scope".
-func containerLeafCgroup(targetID, driver string) string {
-	if driver == cgroupDriverSystemd {
-		return "docker-" + targetID + ".scope"
-	}
-	return targetID
-}
-
-// inspectCgroupParent returns the target container's CgroupParent from inspect.
-// Returns empty string when CgroupParent is not set (standalone Docker defaults)
-// or when inspect fails.
-func (client dockerClient) inspectCgroupParent(ctx context.Context, targetID string) string {
-	inspect, err := client.containerAPI.ContainerInspect(ctx, targetID)
-	if err != nil {
-		log.WithError(err).Warn("failed to inspect target container for cgroup path")
-		return ""
-	}
-	if inspect.HostConfig != nil && inspect.HostConfig.CgroupParent != "" {
-		log.WithField("cgroup-parent", inspect.HostConfig.CgroupParent).Debug("resolved cgroup parent from container inspect")
-		return inspect.HostConfig.CgroupParent
-	}
-	return ""
-}
-
-// defaultCgroupParent returns the default cgroup parent path based on the Docker
-// daemon's cgroup driver when the target container has no explicit CgroupParent set.
-func defaultCgroupParent(targetID, driver string) string {
-	switch driver {
-	case cgroupDriverSystemd:
-		return "system.slice"
-	default:
-		return "/docker/" + targetID
-	}
 }
 
 // stressContainerConfig builds the container and host config for a stress-ng container.
@@ -138,74 +73,8 @@ func stressContainerConfig(targetID string, stressors []string, img, driver, cgr
 		}
 }
 
-// pullImage pulls a Docker image and drains the progress stream.
-func (client dockerClient) pullImage(ctx context.Context, img string) error {
-	log.WithField("img", img).Debug("pulling image")
-	events, err := client.imageAPI.ImagePull(ctx, img, imagetypes.PullOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to pull stress-ng img: %w", err)
-	}
-	defer events.Close()
-	d := json.NewDecoder(events)
-	var pullResponse *imagePullResponse
-	for {
-		if err = d.Decode(&pullResponse); err != nil {
-			if errors.Is(err, io.EOF) {
-				return nil
-			}
-			return fmt.Errorf("failed to decode docker pull result: %w", err)
-		}
-		log.Debug(pullResponse)
-	}
-}
-
-// stressResolveDriver resolves the cgroup driver, parent, and target cgroup path
-// for stress container setup. For default mode, cgroupParent is the resolved path
-// for --cgroup-parent. For inject-cgroup mode, cgroupPath is the target's full
-// cgroup path (if known) to pass as --cgroup-path to cg-inject.
-func (client dockerClient) stressResolveDriver(ctx context.Context, targetID string, injectCgroup bool) (driver, cgroupParent, cgroupPath string, err error) {
-	// resolve the cgroup driver first — needed for correct leaf cgroup naming
-	driver, err = client.cgroupDriver(ctx)
-	if err != nil {
-		// try inspect anyway; if it yields a parent we can still proceed
-		inspectParent := client.inspectCgroupParent(ctx, targetID)
-		if inspectParent == "" {
-			return "", "", "", fmt.Errorf("failed to get docker info: %w", err)
-		}
-		// infer driver from parent path: systemd parents end with .slice
-		if strings.HasSuffix(inspectParent, ".slice") {
-			driver = cgroupDriverSystemd
-		} else {
-			driver = cgroupDriverCgroupfs
-		}
-		cgroupPath = inspectParent + "/" + containerLeafCgroup(targetID, driver)
-	} else {
-		if driver == "" {
-			driver = cgroupDriverCgroupfs
-		}
-		if inspectParent := client.inspectCgroupParent(ctx, targetID); inspectParent != "" {
-			cgroupPath = inspectParent + "/" + containerLeafCgroup(targetID, driver)
-		}
-	}
-
-	if injectCgroup {
-		return driver, cgroupParent, cgroupPath, nil
-	}
-	if cgroupPath == "" {
-		cgroupParent = defaultCgroupParent(targetID, driver)
-		return driver, cgroupParent, cgroupPath, nil
-	}
-	// For default mode, CgroupParent must be a value Docker accepts.
-	// systemd requires a valid slice name (*.slice); cgroupfs accepts any path.
-	if driver == cgroupDriverSystemd {
-		cgroupParent = cgroupPath[:strings.LastIndex(cgroupPath, "/")]
-	} else {
-		cgroupParent = cgroupPath
-	}
-	return driver, cgroupParent, cgroupPath, nil
-}
-
-// execute a stress-ng command in stress-ng Docker container in target container cgroup
+// stressContainerCommand executes a stress-ng command in a stress-ng Docker container
+// in the target container's cgroup.
 func (client dockerClient) stressContainerCommand(ctx context.Context, targetID string, stressors []string, img string, pull, injectCgroup bool) (string, <-chan string, <-chan error, error) {
 	log.WithFields(log.Fields{
 		"target":        targetID,
