+++ pkg/runtime/docker/stress_test.go
@@ -1,14 +1,21 @@
 package docker
 
 import (
+	"bufio"
+	"bytes"
 	"context"
+	"encoding/json"
 	"errors"
+	"io"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/alexei-led/pumba/mocks"
 	ctr "github.com/alexei-led/pumba/pkg/container"
+	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
+	imagetypes "github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/api/types/system"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -495,4 +502,237 @@ func TestStressContainerInjectCgroupCustomImage(t *testing.T) {
 	assert.Equal(t, []string{"/cg-inject"}, []string(capturedConfig.Entrypoint))
 
 	api.AssertExpectations(t)
+}
+
+func Test_dockerClient_stressContainerCommand(t *testing.T) {
+	type args struct {
+		ctx       context.Context
+		targetID  string
+		stressors []string
+		image     string
+		pull      bool
+	}
+	tests := []struct {
+		name       string
+		args       args
+		mockInit   func(context.Context, *mocks.APIClient, *mockConn, string, []string, string, bool)
+		want       string
+		wantOutput string
+		wantErr    bool
+		wantErrCh  bool
+	}{
+		{
+			name: "stress test with pull image",
+			args: args{
+				ctx:       context.TODO(),
+				targetID:  "123",
+				stressors: []string{"--cpu", "4"},
+				image:     "test/stress-ng",
+				pull:      true,
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				pullResponse := imagePullResponse{
+					Status:   "ok",
+					Error:    "no error",
+					Progress: "done",
+					ProgressDetail: struct {
+						Current int `json:"current"`
+						Total   int `json:"total"`
+					}{
+						Current: 100,
+						Total:   100,
+					},
+				}
+				pullResponseByte, _ := json.Marshal(pullResponse)
+				readerResponse := bytes.NewReader(pullResponseByte)
+				engine.EXPECT().ImagePull(ctx, image, imagetypes.PullOptions{}).Return(io.NopCloser(readerResponse), nil)
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{
+					Conn:   conn,
+					Reader: bufio.NewReader(strings.NewReader("stress completed")),
+				}, nil)
+				engine.EXPECT().ContainerStart(ctx, "000", mock.Anything).Return(nil)
+				inspect := container.InspectResponse{
+					ContainerJSONBase: &container.ContainerJSONBase{
+						State: &container.State{},
+					},
+				}
+				conn.On("Close").Return(nil)
+				engine.EXPECT().ContainerInspect(ctx, "000").Return(inspect, nil)
+			},
+			want:       "000",
+			wantOutput: "stress completed",
+		},
+		{
+			name: "stress test fail to pull image",
+			args: args{
+				ctx:   context.TODO(),
+				image: "test/stress-ng",
+				pull:  true,
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ImagePull(ctx, image, imagetypes.PullOptions{}).Return(io.NopCloser(bytes.NewReader([]byte("{}"))), errors.New("failed to pull image"))
+			},
+			wantErr: true,
+		},
+		{
+			name: "stress test without pull image",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{
+					Conn:   conn,
+					Reader: bufio.NewReader(strings.NewReader("stress completed")),
+				}, nil)
+				engine.EXPECT().ContainerStart(ctx, "000", mock.Anything).Return(nil)
+				inspect := container.InspectResponse{
+					ContainerJSONBase: &container.ContainerJSONBase{
+						State: &container.State{},
+					},
+				}
+				engine.EXPECT().ContainerInspect(ctx, "000").Return(inspect, nil)
+				conn.On("Close").Return(nil)
+			},
+			want:       "000",
+			wantOutput: "stress completed",
+		},
+		{
+			name: "stress-ng exit with error",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{
+					Conn:   conn,
+					Reader: bufio.NewReader(strings.NewReader("stress completed")),
+				}, nil)
+				engine.EXPECT().ContainerStart(ctx, "000", mock.Anything).Return(nil)
+				inspect := container.InspectResponse{
+					ContainerJSONBase: &container.ContainerJSONBase{
+						State: &container.State{ExitCode: 1},
+					},
+				}
+				conn.On("Close").Return(nil)
+				engine.EXPECT().ContainerInspect(ctx, "000").Return(inspect, nil)
+			},
+			want:      "000",
+			wantErrCh: true,
+		},
+		{
+			name: "fail to inspect stress-ng container",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{
+					Conn:   conn,
+					Reader: bufio.NewReader(strings.NewReader("stress completed")),
+				}, nil)
+				engine.EXPECT().ContainerStart(ctx, "000", mock.Anything).Return(nil)
+				conn.On("Close").Return(nil)
+				engine.EXPECT().ContainerInspect(ctx, "000").Return(container.InspectResponse{}, errors.New("filed to inspect"))
+			},
+			want:       "000",
+			wantOutput: "stress completed",
+			wantErrCh:  true,
+		},
+		{
+			name: "fail to start stress-ng container",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{
+					Conn:   conn,
+					Reader: bufio.NewReader(strings.NewReader("stress completed")),
+				}, nil)
+				engine.EXPECT().ContainerStart(ctx, "000", mock.Anything).Return(errors.New("failed to start"))
+				conn.On("Close").Return(nil).Maybe()
+				inspect := container.InspectResponse{
+					ContainerJSONBase: &container.ContainerJSONBase{
+						State: &container.State{},
+					},
+				}
+				engine.EXPECT().ContainerInspect(ctx, "000").Return(inspect, nil).Maybe()
+			},
+			want:    "000",
+			wantErr: true,
+		},
+		{
+			name: "fail to attach to stress-ng container",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{ID: "000"}, nil)
+				engine.EXPECT().ContainerAttach(ctx, "000", mock.Anything).Return(types.HijackedResponse{}, errors.New("failed to attach"))
+			},
+			wantErr: true,
+		},
+		{
+			name: "fail to create to stress-ng container",
+			args: args{
+				ctx: context.TODO(),
+			},
+			mockInit: func(ctx context.Context, engine *mocks.APIClient, conn *mockConn, targetID string, stressors []string, image string, pool bool) {
+				engine.EXPECT().Info(mock.Anything).Return(system.Info{CgroupDriver: "cgroupfs"}, nil)
+				engine.EXPECT().ContainerInspect(mock.Anything, targetID).Return(DetailsResponse(AsMap("ID", targetID)), nil).Once()
+				engine.EXPECT().ContainerCreate(ctx, mock.Anything, mock.Anything, mock.Anything, mock.Anything, "").Return(container.CreateResponse{}, errors.New("failed to create"))
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			mockClient := NewMockEngine()
+			mConn := &mockConn{}
+			client := dockerClient{
+				containerAPI: mockClient,
+				imageAPI:     mockClient,
+				systemAPI:    mockClient,
+			}
+			tt.mockInit(tt.args.ctx, mockClient, mConn, tt.args.targetID, tt.args.stressors, tt.args.image, tt.args.pull)
+			got, got1, got2, err := client.stressContainerCommand(tt.args.ctx, tt.args.targetID, tt.args.stressors, tt.args.image, tt.args.pull, false)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("dockerClient.stressContainerCommand() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("dockerClient.stressContainerCommand() got = %v, want %v", got, tt.want)
+			}
+			if err == nil && (got1 != nil || got2 != nil) {
+				select {
+				case output := <-got1:
+					if output != tt.wantOutput {
+						t.Errorf("dockerClient.stressContainerCommand() got = %v, from output channel, want %v", output, tt.wantOutput)
+					}
+				case err = <-got2:
+					if (err != nil) != tt.wantErrCh {
+						t.Errorf("dockerClient.stressContainerCommand() error in error channel = %v, wantErrCh %v", err, tt.wantErrCh)
+					}
+				}
+			}
+			mockClient.AssertExpectations(t)
+			mConn.AssertExpectations(t)
+		})
+	}
 }
